feat(Core): Accepted Types are now defined in the route
In the previous implementation, the accepted Content-Types were
specified when reading the body in the handler and the 415 had to be
manually sent. This is now not the case. The accepted types are defined
in the router definition and the 415 error, as well as the body type
parsing are completely transparent in the handler.